+++ src/firmware/quake/net_main.c
@@ -46,7 +46,7 @@ static qboolean	listening = false;
 qboolean	slistInProgress = false;
 qboolean	slistSilent = false;
 qboolean	slistLocal = true;
-static double	slistStartTime;
+static float	slistStartTime;
 static int		slistLastShown;
 
 static void Slist_Send(void);
@@ -90,9 +90,9 @@ qboolean recording = false;
 int	net_driverlevel;
 
 
-double			net_time;
+float			net_time;
 
-double SetNetTime(void)
+float SetNetTime(void)
 {
 	net_time = Sys_FloatTime();
 	return net_time;
@@ -455,7 +455,7 @@ NET_CheckNewConnections
 
 struct
 {
-	double	time;
+	float	time;
 	int		op;
 	long	session;
 } vcrConnect;
@@ -549,7 +549,7 @@ returns -1 if connection is invalid
 
 struct
 {
-	double	time;
+	float	time;
 	int		op;
 	long	session;
 	int		ret;
@@ -637,7 +637,7 @@ returns -1 if the connection died
 */
 struct
 {
-	double	time;
+	float	time;
 	int		op;
 	long	session;
 	int		r;
@@ -742,7 +742,7 @@ qboolean NET_CanSendMessage (qsocket_t *sock)
 
 int NET_SendToAll(sizebuf_t *data, int blocktime)
 {
-	double		start;
+	float		start;
 	int			i;
 	int			count = 0;
 	qboolean	state1 [MAX_SCOREBOARD];
@@ -976,7 +976,7 @@ void NET_Poll(void)
 }
 
 
-void SchedulePollProcedure(PollProcedure *proc, double timeOffset)
+void SchedulePollProcedure(PollProcedure *proc, float timeOffset)
 {
 	PollProcedure *pp, *prev;
