+++ src/firmware/quake/sv_phys.c
@@ -932,8 +932,8 @@ int SV_TryUnstick (edict_t *ent, vec3_t oldvel)
 		ent->v. velocity[2] = 0;
 		clip = SV_FlyMove (ent, 0.1, &steptrace);
 
-		if ( fabs(oldorg[1] - ent->v.origin[1]) > 4
-		|| fabs(oldorg[0] - ent->v.origin[0]) > 4 )
+		if ( fabsf(oldorg[1] - ent->v.origin[1]) > 4
+		|| fabsf(oldorg[0] - ent->v.origin[0]) > 4 )
 		{
 //Con_DPrintf ("unstuck!\n");
 			return clip;
@@ -1016,8 +1016,8 @@ void SV_WalkMove (edict_t *ent)
 // in the clipping hulls
 	if (clip)
 	{
-		if ( fabs(oldorg[1] - ent->v.origin[1]) < 0.03125
-		&& fabs(oldorg[0] - ent->v.origin[0]) < 0.03125 )
+		if ( fabsf(oldorg[1] - ent->v.origin[1]) < 0.03125
+		&& fabsf(oldorg[0] - ent->v.origin[0]) < 0.03125 )
 		{	// stepping up didn't make any progress
 			clip = SV_TryUnstick (ent, oldvel);
 		}
@@ -1408,7 +1408,7 @@ void SV_Physics_Step (edict_t *ent)
 			if (!(ent->v.health <= 0.0 && !SV_CheckBottom(ent)))
 			{
 				vel = ent->v.velocity;
-				speed = sqrt(vel[0]*vel[0] +vel[1]*vel[1]);
+				speed = sqrtf(vel[0]*vel[0] +vel[1]*vel[1]);
 				if (speed)
 				{
 					friction = sv_friction.value;
@@ -1571,7 +1571,7 @@ trace_t SV_Trace_Toss (edict_t *ent, edict_t *ignore)
 	trace_t	trace;
 	vec3_t	move;
 	vec3_t	end;
-	double	save_frametime;
+	float	save_frametime;
 //	extern particle_t	*active_particles, *free_particles;
 //	particle_t	*p;
